# module: update cespare/xxhash to v2