--- a/bigcache.go
+++ b/bigcache.go
@@ -4,7 +4,7 @@ import (
 	"sync"
 	"sync/atomic"
 
-	"github.com/cespare/xxhash"
+	xxhash "github.com/cespare/xxhash/v2"
 )
 
 // maxSubvalueLen is the maximum size of subvalue chunk.