--- a/go.mod
+++ b/go.mod
@@ -1,9 +1,11 @@
 module github.com/VictoriaMetrics/fastcache
 
 require (
+	github.com/OneOfOne/xxhash v1.2.5 // indirect
 	github.com/allegro/bigcache v1.1.1-0.20190116153254-84a0ff3f153c
-	github.com/cespare/xxhash v1.1.0
+	github.com/cespare/xxhash/v2 v2.0.1-0.20190104013014-3767db7a7e18
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/golang/snappy v0.0.0-20180518054509-2e65f85255db
+	github.com/spaolacci/murmur3 v1.1.0 // indirect
 	github.com/stretchr/testify v1.3.0 // indirect
 )