--- a/go.mod
+++ b/go.mod
@@ -2,10 +2,10 @@ module github.com/VictoriaMetrics/fastcache
 
 require (
 	github.com/OneOfOne/xxhash v1.2.5 // indirect
-	github.com/allegro/bigcache v1.1.1-0.20190116153254-84a0ff3f153c
+	github.com/allegro/bigcache v1.2.1-0.20190218064605-e24eb225f156
 	github.com/cespare/xxhash/v2 v2.0.1-0.20190104013014-3767db7a7e18
 	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/golang/snappy v0.0.0-20180518054509-2e65f85255db
-	github.com/spaolacci/murmur3 v1.1.0 // indirect
+	github.com/golang/snappy v0.0.1
+	github.com/spaolacci/murmur3 v1.0.1-0.20190317074736-539464a789e9 // indirect
 	github.com/stretchr/testify v1.3.0 // indirect
 )