@narqo 2019-05-25
Wasn't this a breaking change? For projects that doesn't use gomodules, e.g. [dep][], there is an error when one tries to use `fastcache@1.5.0`

```
Solving failure: No versions of github.com/cespare/xxhash met constraints:
	v2.0.0: Could not introduce github.com/cespare/xxhash@v2.0.0, as its subpackage github.com/cespare/xxhash/v2 is missing. (Package is required by github.com/VictoriaMetrics/fastcache@v1.5.0.)
```

See https://github.com/golang/dep/issues/2139

[dep]: https://github.com/golang/dep