@valyala 2019-05-26
@narqo , yup, this is unfortunate. There are the following solutions:
- to stick to [v1.4.6](https://github.com/VictoriaMetrics/fastcache/tree/v1.4.6) when working with `dep`;
- to wait for the fix for https://github.com/golang/dep/issues/2139;
- to upgrade to `go mod`, since `dep` is deprecated.

I'd prefer the last one :)