@zchee 2019-05-27
@narqo @valyala
Ohh, sorry I forget this change is no backward compatibility.
But yes, I also suggest @valyala solution (last one)